+++ configs/gpt-sovits_template.toml
@@ -1,0 +1,57 @@
+[pipeline]
+# pipeline预设配置
+default_preset = "default"  # 默认使用的预设名称
+
+# pipeline平台预设映射
+[pipeline.platform_presets]
+qq = "default"       # QQ平台使用default预设
+# nonebot-qq = "default"
+
+[tts]
+# GPT-SoVITS API 配置
+host = "127.0.0.1"
+port = 9880
+
+# 参考音频配置
+ref_audio_path = "zh-CN-XiaoyiNeural.mp3"    # 默认参考音频路径
+prompt_text = "这是一段测试音频"            # 默认提示文本
+aux_ref_audio_paths = []                    # 辅助参考音频路径列表
+
+# 语音合成基础配置
+text_language = "zh"      # 文本语言
+prompt_language = "zh"    # 提示语言
+media_type = "wav"        # 音频格式:wav/raw/ogg/aac
+
+# GPT-SoVITS 模型参数
+top_k = 12                # top k 采样
+top_p = 1.0              # top p 采样
+temperature = 1.0        # 温度系数
+batch_size = 1           # 批处理大小
+batch_threshold = 0.75   # 批处理阈值
+speed_factor = 1.0       # 语速控制
+text_split_method = "cut5" # 文本分割方法
+repetition_penalty = 1.35 # 重复惩罚系数
+sample_steps = 32        # VITS采样步数
+super_sampling = false    # 是否启用超采样
+
+[tts.models]
+# GPT模型配置
+gpt_model = "GPT_SoVITS/pretrained_models/XiangNaimei-e5.ckpt"  # GPT模型路径
+sovits_model = "GPT_SoVITS/pretrained_models/XiangNaimei_e4_s248.pth"  # SoVITS模型路径
+
+# 自定义角色预设
+[tts.models.presets]
+# 预设角色配置示例
+[tts.models.presets.default]
+name = "默认角色"
+ref_audio = "GPT_SoVITS/pretrained_models/voices/平淡.WAV"
+prompt_text = "嗨！我每天都会像这样做发声训练的。"
+gpt_model = ""     # 为空则使用默认GPT模型
+sovits_model = ""  # 为空则使用默认SoVITS模型
+
+[tts.models.presets.custom1]
+name = "kanami"
+ref_audio = "GPT_SoVITS/pretrained_models/voices/平淡.WAV"
+prompt_text = "嗨！我每天都会像这样做发声训练的。"
+gpt_model = "GPT_SoVITS/pretrained_models/XiangNaimei-e5.ckpt"
+sovits_model = "GPT_SoVITS/pretrained_models/XiangNaimei_e4_s248.pth"
