+++ src/plugins/GPT_Sovits/__init__.py
@@ -1,0 +1,3 @@
+from .tts_model import TTSModel 
+
+__all__ = ["TTSModel"]
